better e2e
